requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
charset-normalizer==3.0.1
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.1
daphne==3.0.2
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.13.0
django-lookup-extensions==0.2.1
django-ninja==0.21.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
executing==1.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
install==1.3.5
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
Pygments==2.14.0
pyOpenSSL==23.1.0
pytz==2022.7.1
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.2.6
zope.interface==6.0
